CODERUB: Fixed To Use Partials
