+++ InternTrack.Portal.Web/Brokers/Apis/ApiBroker.cs
@@ -1,0 +1,49 @@
+﻿// -------------------------------------------------------
+// Copyright (c) Coalition of the Good-Hearted Engineers
+// FREE TO USE FOR THE WORLD
+// -------------------------------------------------------
+
+using System;
+using System.Net.Http;
+using System.Threading.Tasks;
+using InternTrack.Portal.Web.Models.Configurations;
+using Microsoft.Extensions.Configuration;
+using RESTFulSense.Clients;
+
+namespace InternTrack.Portal.Web.Brokers.Apis
+{
+    public partial class ApiBroker : IApiBroker
+    {
+        private readonly IRESTFulApiFactoryClient apiClient;
+        private readonly HttpClient httpClient;
+
+        public ApiBroker(HttpClient httpClient, IConfiguration configuration)
+        {
+            this.httpClient = httpClient;
+            this.apiClient = GetApiClient(configuration);
+        }
+
+        private async ValueTask<T> GetAsync<T>(string relativeUrl) =>
+             await this.apiClient.GetContentAsync<T>(relativeUrl);
+
+        private async ValueTask<T> PostAsync<T>(string relativeUrl, T content) =>
+            await this.apiClient.PostContentAsync<T>(relativeUrl, content);
+
+        private async ValueTask<T> PutAsync<T>(string relativeUrl, T content) =>
+            await this.apiClient.PutContentAsync<T>(relativeUrl, content);
+
+        private async ValueTask<T> DeleteAsync<T>(string relativeUrl) =>
+            await this.apiClient.DeleteContentAsync<T>(relativeUrl);
+
+        private IRESTFulApiFactoryClient GetApiClient(IConfiguration configuration)
+        {
+            LocalConfigurations localConfigurations =
+            configuration.Get<LocalConfigurations>();
+
+            string apiBaseUrl = localConfigurations.ApiConfigurations.Url;
+
+            return new RESTFulApiFactoryClient(this.httpClient);
+            throw new NotImplementedException();
+        }
+    }
+}
